05-ai-copilots-and-agents-theory/requirements.txt
streamlit==1.39.0
langchain==0.3.2
langchain-anthropic==0.2.1
python-dotenv==1.0.1
requests==2.32.3
